Fix(main.ts): segregated all main.ts content into respective files and folders
