+++ simon/src/js/utils/delay.ts
@@ -1,0 +1,4 @@
+// Functions
+export const delay = (ms: number) => {
+    return new Promise(resolve => setTimeout(resolve, ms))
+}
