+++ simon/src/js/utils/dom.ts
@@ -1,0 +1,2 @@
+// Currently empty, in any future iterations place dom element
+// references and helpers here
